adding functionality
Adding some functionality and working through the logic of reading in, writing out, storing strings, converting types, and editing strings and lists. This update adds a lot to the inteded flow and some of the basic maths needed for converting user input to the expected format of the Marlin firmware.